update script and compose setup for xai mainnet node
