--- infrastructure/nitro-node-wrapper/.env
@@ -1,1 +1,0 @@
-ARBITRUM_NODE_IMAGE_TAG=v2.1.0-72ccc0c
