+++ infrastructure/nitro-node-wrapper/docker-compose.yml
@@ -1,23 +1,52 @@
 version: '3'
 
 services:
-  your_service_name:
-    image: offchainlabs/nitro-node:${ARBITRUM_NODE_IMAGE_TAG}
+  nitro-node:
+    container_name: arb1-nitro-node
+    image: offchainlabs/nitro-node:v2.2.2-8f33fea
+    command:
+      --init.url="https://snapshot.arbitrum.foundation/arb1/nitro-pruned.tar"
+      --parent-chain.connection.url=${ETH_RPC_URL}
+      --chain.id=42161
+      --http.api=net,web3,eth
+      --http.corsdomain=*
+      --http.addr=0.0.0.0
+      --http.vhosts=*
+    volumes:
+      - /opt/arbitrum:/home/user/.arbitrum
+    ports:
+      - "8547:8547" # This should not be needed for connection between containers
+    restart: unless-stopped
+    dns:
+      - 8.8.8.8
+      - 8.8.4.4
+
+  xai-node:
+    image: offchainlabs/nitro-node:v2.2.2-8f33fea
     container_name: xai-public-node
     command:
-      - --init.url=file:///home/user/nitro-pruned.tar
-      - --conf.file=/home/user/.arbitrum/config.json
+      - --conf.file=/home/user/node-config/config.json
       - --metrics
       - --ws.port=8548
       - --ws.addr=0.0.0.0
       - --ws.origins=*
-    tty: false
-    ports:
-      - "8547:8547"
-      - "8548:8548"
-      - "9642:9642"
-    # restart: always
     volumes:
-      - ./arbitrum:/home/user/.arbitrum
-      - S:\arb1-node-data:/home/user/data:delegated
-      - D:\ARBITRUM_TMP_DOWNLOAD\nitro-pruned.tar:/home/user/nitro-pruned.tar
+      - /opt/node/build/xai-mainnet.config.json:/home/user/node-config/config.json
+      - /opt/xai:/home/user/.arbitrum
+    depends_on:
+      - nitro-node
+    restart: unless-stopped
+
+  publisher:
+    image: cryptitaustria/xai-public-node-publisher:latest
+    container_name: js-publisher
+    environment:
+      - PROJECT_ID=${PROJECT_ID}
+      - SERVICE_ACCOUNT_EMAIL=${SERVICE_ACCOUNT_EMAIL}
+      - SERVICE_ACCOUNT_PRIVATE_KEY=${SERVICE_ACCOUNT_PRIVATE_KEY}
+      - BUCKET_NAME=${BUCKET_NAME}
+    volumes:
+      - /var/run/docker.sock:/var/run/docker.sock
+    depends_on:
+      - xai-node
+    restart: unless-stopped
