+++ infrastructure/nitro-node-wrapper/publisher/package.json
