+++ infrastructure/nitro-node-wrapper/sample.env
@@ -1,0 +1,12 @@
+# Arb1 Node Settings
+# The Etehreum mainnet RPC to sync the ARB1 node with (prob > 100K requests to sync)
+ETH_RPC_URL=
+
+# GBUCKET Project ID
+PROJECT_ID=
+# GBUCKET Account Email
+SERVICE_ACCOUNT_EMAIL=
+# GBUCKET PRIVATE KEY
+SERVICE_ACCOUNT_PRIVATE_KEY=
+# GBUCKET BUCKET NAME
+BUCKET_NAME=
